requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
streamlit>=1.31.0
